Move binary like formatting to its own file
